+++ packages/preview3-shim/lib/nodejs/workers/cli-worker.js
