+++ packages/preview3-shim/lib/nodejs/workers/filesystem-worker.js
