+++ packages/preview3-shim/lib/nodejs/workers/resource-worker.js
